исправлял ошибки наптсал исключения

+++ main-service/src/main/java/ru/practicum/exploreWithMe/stats/request/dto/EventRequestStatusUpdateRequest.java
@@ -2,7 +2,7 @@ package ru.practicum.exploreWithMe.stats.request.dto;
 
 import lombok.Builder;
 import lombok.Data;
-import ru.practicum.exploreWithMe.stats.events.model.Status;
+import ru.practicum.exploreWithMe.stats.statuses.Status;
 
 
 import java.util.List;
@@ -10,7 +10,7 @@ import java.util.List;
 @Data
 @Builder
 public class EventRequestStatusUpdateRequest {
-    private List<Integer> requestIds;
+    private List<Long> requestIds;
     private Status status;
 
 }
